requirements.txt
PyPDF2
stix2
pyyaml
requests
openai
